tests: Move progress recording coverage
The command layer now imports progress helpers from data.progress, while
the test suite still keeps hunk recording checks in the hunk navigation
test module.

The project does not keep validation aligned with the module that owns
progress storage. That makes hunk_tracking look like a supported entry
point for progress behavior.

This commit addresses that by moving hunk recording assertions into
test_progress and removing progress helper imports from the hunk navigation
tests.

The next commit will remove the progress helper re-exports from
hunk_tracking.
